requirements.txt
flask==2.3.3
ultralytics==8.0.200
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0
